Ensure compatibility with ECMAScript 2015 modules
The module evaluation context of ECMAScript 2015 modules is not the
global context. As such `this` does not point to the global `window`
object. This is problematic for users of ECMAScript 2015 modules as the
included `window.performance.now` polyfill will fail to work.

Instead of `this`, the polyfill should be added to the global `window`
object directly. This is unproblematic because the following code
assumes that `performance.now` will be available under the global
`window` object, i.e. the code was never execution environment agnostic.

Furthermore, established polyfills like the one from Paul Irish also
directly register the polyfill on the global `window` object.

Fixes #209

[1] https://gist.github.com/paulirish/5438650